Move cmd configschema
