+++ cmd/configschema/common_test.go
