+++ cmd/configschema/configs.go
