+++ cmd/configschema/docsgen/README.md
