+++ cmd/configschema/docsgen/docsgen/cli.go
@@ -24,8 +24,9 @@ import (
 	"strings"
 	"text/template"
 
-	"go.opentelemetry.io/collector/cmd/configschema"
 	"go.opentelemetry.io/collector/component"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/configschema"
 )
 
 const mdFileName = "config.md"
