+++ cmd/configschema/docsgen/docsgen/template.go
