+++ cmd/configschema/docsgen/docsgen/testdata/otlp-receiver.json
