+++ cmd/configschema/docsgen/docsgen/testdata/test.tmpl
