+++ cmd/configschema/fields_test.go
