+++ cmd/configschema/resolver_test.go
